Move response hooks setup and handling out of RESTClient
